Add preseed input for action
Add the preseed input for the action, making it possible to fully customise how
LXD is initialised.

Update the README to include documentation for this input.

Add a new test which uses this new input with lxd init. The preseed for this
test was obtained by running lxd init interactively on a fresh noble install and
printing the preseed at the end.

Signed-off-by: Alexis Janon <alexis.janon@canonical.com>